move ../java/../dl/ModelPlexConnector.scala -> ../scala/../dl/ModelPlexConnector.scala
